Add ability to plot multiple traces with same trace offset. Change color to a pointer for traces

+++ source/behaviortrackerworker.h
@@ -28,7 +28,7 @@ public:
     void setUpDLCLive();
     QVector<float> getDLCLivePose(cv::Mat frame);
     void setParameters(QString name, cv::Mat *frameBuf, int bufSize, QAtomicInt *acqFrameNum);
-    void setPoseBufferParameters(QVector<float> *poseBuf, int *poseFrameNumBuf, int poseBufSize, QAtomicInt *btPoseFrameNum, QSemaphore *free, QSemaphore *used, uint8_t *pColors);
+    void setPoseBufferParameters(QVector<float> *poseBuf, int *poseFrameNumBuf, int poseBufSize, QAtomicInt *btPoseFrameNum, QSemaphore *free, QSemaphore *used, float *pColors);
     void getColors();
 
 signals:
@@ -61,7 +61,7 @@ private:
     QVector<float> *poseBuffer;
     int *poseFrameNumBuffer;
     QAtomicInt *m_btPoseCount;
-    uint8_t *colors;
+    float *colors;
 
     // For DLC python class
     PyObject *pInstance;
